+++ queue/src/fiskaltrust.Middleware.Localization.QueueGR/Models/Cases/ReceiptCaseFlagsGR.cs
@@ -7,7 +7,6 @@ public enum ReceiptCaseFlagsGR : long
     IsSelfPricingOperation = 0x0100_0000_0000,
 }
 
-
 public static class ReceiptCaseFlagsGRExt
 {
     public static ReceiptCase WithFlag(this ReceiptCase self, ReceiptCaseFlagsGR flag) => (ReceiptCase) ((long) self | (long) flag);
