+++ queue/src/fiskaltrust.Middleware.Localization.QueueGR/QueueGRBootstrapper.cs
@@ -1,6 +1,6 @@
 ﻿using System.Text.Json;
-using fiskaltrust.Middleware.Localization.QueueGR.GRSSCD.myDataSCU;
 using fiskaltrust.Middleware.Localization.QueueGR.Processors;
+using fiskaltrust.Middleware.Localization.QueueGR.SCU.GR.MyData;
 using fiskaltrust.Middleware.Localization.v2;
 using fiskaltrust.Middleware.Localization.v2.Configuration;
 using fiskaltrust.Middleware.Localization.v2.Interface;
