+++ queue/src/fiskaltrust.Middleware.Localization.QueueGR/SCU/GR/MyData/MyDataAADEValidation.cs
@@ -3,7 +3,7 @@ using fiskaltrust.Middleware.Localization.QueueGR.Models.Cases;
 using fiskaltrust.Middleware.Localization.v2.Helpers;
 using fiskaltrust.Middleware.Localization.v2.Models.ifPOS.v2.Cases;
 
-namespace fiskaltrust.Middleware.Localization.QueueGR.GRSSCD.AADE;
+namespace fiskaltrust.Middleware.Localization.QueueGR.SCU.GR.MyData;
 
 public class MyDataAADEValidation
 {
