+++ queue/src/fiskaltrust.Middleware.Localization.QueueGR/SCU/GR/MyData/MyDataApiClient.cs
@@ -1,19 +1,15 @@
-﻿using System.Net.Http.Headers;
-using System.Security.Cryptography;
-using System.Text;
+﻿using System.Text;
 using System.Text.Json;
 using System.Xml.Serialization;
 using fiskaltrust.Api.POS.Models.ifPOS.v2;
-using fiskaltrust.Middleware.Localization.QueueGR.GRSSCD.AADE;
-using fiskaltrust.Middleware.Localization.QueueGR.Interface;
+using fiskaltrust.Middleware.Localization.QueueGR.GRSSCD;
 using fiskaltrust.Middleware.Localization.QueueGR.Models.Cases;
 using fiskaltrust.Middleware.Localization.v2.Interface;
 using fiskaltrust.Middleware.Localization.v2.Models.ifPOS.v2.Cases;
 using fiskaltrust.storage.V0.MasterData;
-using Org.BouncyCastle.Asn1.Ocsp;
 
 #pragma warning disable
-namespace fiskaltrust.Middleware.Localization.QueueGR.GRSSCD.myDataSCU;
+namespace fiskaltrust.Middleware.Localization.QueueGR.SCU.GR.MyData;
 
 public class MyDataApiClient : IGRSSCD
 {
