+++ queue/src/fiskaltrust.Middleware.Localization.QueueGR/SCU/GR/MyData/ReceiptRequestExtensions.cs
@@ -1,7 +1,7 @@
 ﻿using fiskaltrust.Api.POS.Models.ifPOS.v2;
 using fiskaltrust.Middleware.Localization.v2.Models.ifPOS.v2.Cases;
 
-namespace fiskaltrust.Middleware.Localization.QueueGR.GRSSCD.AADE;
+namespace fiskaltrust.Middleware.Localization.QueueGR.SCU.GR.MyData;
 
 public static class ReceiptRequestExtensions
 {
