+++ queue/src/fiskaltrust.Middleware.Localization.QueueGR/fiskaltrust.Middleware.Localization.QueueGR.csproj
@@ -17,5 +17,32 @@
         <ProjectReference Include="..\fiskaltrust.Middleware.Storage.AzureTableStorage\fiskaltrust.Middleware.Storage.AzureTableStorage.csproj" />
         <ProjectReference Include="..\fiskaltrust.Middleware.Storage\fiskaltrust.Middleware.Storage.csproj" />
     </ItemGroup>
+
+    <ItemGroup>
+      <None Update="SCU\GR\MyData\Models\expensesClassification-v1.0.9.xsd">
+        <SubType>Designer</SubType>
+      </None>
+      <None Update="SCU\GR\MyData\Models\incomeClassification-v1.0.9.xsd">
+        <SubType>Designer</SubType>
+      </None>
+      <None Update="SCU\GR\MyData\Models\InvoicesDoc-v1.0.9.xsd">
+        <SubType>Designer</SubType>
+      </None>
+      <None Update="SCU\GR\MyData\Models\paymentMethods-v1.0.9.xsd">
+        <SubType>Designer</SubType>
+      </None>
+      <None Update="SCU\GR\MyData\Models\requestedInvoicesDoc-v1.0.9.xsd">
+        <SubType>Designer</SubType>
+      </None>
+      <None Update="SCU\GR\MyData\Models\RequestedProviderDoc-v1.0.9.xsd">
+        <SubType>Designer</SubType>
+      </None>
+      <None Update="SCU\GR\MyData\Models\RequestVatInfoResponse-v1.0.9.xsd">
+        <SubType>Designer</SubType>
+      </None>
+      <None Update="SCU\GR\MyData\Models\response-v1.0.9.xsd">
+        <SubType>Designer</SubType>
+      </None>
+    </ItemGroup>
     
 </Project>
