+++ queue/test/fiskaltrust.Middleware.Localization.QueueGR.IntegrationTest/QueueGR/AADECertificationTestsSelfPricing.cs
@@ -4,7 +4,7 @@ using System.Text;
 using System.Text.Json;
 using System.Xml.Serialization;
 using fiskaltrust.Api.POS.Models.ifPOS.v2;
-using fiskaltrust.Middleware.Localization.QueueGR.GRSSCD.AADE;
+using fiskaltrust.Middleware.Localization.QueueGR.SCU.GR.MyData;
 using fiskaltrust.Middleware.Localization.v2.Configuration;
 using fiskaltrust.Middleware.Localization.v2.Models.ifPOS.v2.Cases;
 using FluentAssertions;
