+++ queue/test/fiskaltrust.Middleware.Localization.QueueGR.IntegrationTest/QueueGR/IGRSSCD/MyDataApiTests.cs
@@ -1,6 +1,6 @@
 ﻿using System.Text.Json;
 using fiskaltrust.Api.POS.Models.ifPOS.v2;
-using fiskaltrust.Middleware.Localization.QueueGR.GRSSCD.myDataSCU;
+using fiskaltrust.Middleware.Localization.QueueGR.SCU.GR.MyData;
 using fiskaltrust.Middleware.Localization.v2.Models.ifPOS.v2.Cases;
 using Xunit;
