+++ queue/test/fiskaltrust.Middleware.Localization.QueueGR.IntegrationTest/ReceiptExamples.cs
@@ -1,7 +1,5 @@
 ﻿using fiskaltrust.Api.POS.Models.ifPOS.v2;
-using fiskaltrust.Middleware.Localization.QueueGR.GRSSCD.AADE;
-using fiskaltrust.Middleware.Localization.QueueGR.GRSSCD.AADE.Models;
-using fiskaltrust.Middleware.Localization.QueueGR.UnitTest;
+using fiskaltrust.Middleware.Localization.QueueGR.SCU.GR.MyData.Models;
 using fiskaltrust.Middleware.Localization.v2.Models;
 using fiskaltrust.Middleware.Localization.v2.Models.ifPOS.v2.Cases;
